extract comparison_requirements from ConditionExampleGenerator
